+++ parse.c
@@ -103,7 +103,7 @@ Type *find_typedef(Token *tok) {
   return NULL;
 }
 
-Node *new_num(int val, Token *tok) {
+Node *new_num(long val, Token *tok) {
   Node *node = new_node(ND_NUM, tok);
   node->val = val;
   return node;
